requirements.txt
asgiref==3.8.1
awscli==1.36.6
boto3==1.35.65
botocore==1.35.65
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.2
docutils==0.16
gunicorn==23.0.0
jmespath==1.0.1
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pyasn1==0.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.3
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2
